Port DeepCopy rank 1, 2 & 3 tests
- port remaining DeepCopy perf tests
- use existing header for DeepCopy benchmark helpers
