+++ core/perf_test/CMakeLists.txt
@@ -22,10 +22,6 @@ SET(SOURCES
   PerfTestHexGrad.cpp
   PerfTest_CustomReduction.cpp
   PerfTest_ExecSpacePartitioning.cpp
-  PerfTest_ViewCopy_a123.cpp
-  PerfTest_ViewCopy_b123.cpp
-  PerfTest_ViewCopy_c123.cpp
-  PerfTest_ViewCopy_d123.cpp
   PerfTest_ViewAllocate.cpp
   PerfTest_ViewFill_123.cpp
   PerfTest_ViewFill_45.cpp
@@ -186,7 +182,11 @@ ENDFUNCTION()
 SET(
   BENCHMARK_SOURCES
   BenchmarkMain.cpp
-  Benchmark_ViewCopy.cpp
+  PerfTest_ViewCopy.cpp
+  PerfTest_ViewCopy_a123.cpp
+  PerfTest_ViewCopy_b123.cpp
+  PerfTest_ViewCopy_c123.cpp
+  PerfTest_ViewCopy_d123.cpp
   PerfTest_ViewCopy_a45.cpp
   PerfTest_ViewCopy_b45.cpp
   PerfTest_ViewCopy_c45.cpp
