+++ core/perf_test/PerfTest_ViewCopy.hpp
@@ -42,71 +42,142 @@
 //@HEADER
 */
 
+#ifndef KOKKOS_CORE_PERFTEST_BENCHMARK_VIEW_COPY_HPP
+#define KOKKOS_CORE_PERFTEST_BENCHMARK_VIEW_COPY_HPP
+
 #include <Kokkos_Core.hpp>
-#include <gtest/gtest.h>
-#include <cstdio>
-#include <PerfTest_Category.hpp>
+#include <benchmark/benchmark.h>
+#include <cmath>
 
 namespace Test {
 
+void report_results(benchmark::State& state, double time);
+
 template <class ViewTypeA, class ViewTypeB>
-double deepcopy_view(ViewTypeA& a, ViewTypeB& b, int repeat) {
-  Kokkos::Timer timer;
-  for (int i = 0; i < repeat; i++) {
-    Kokkos::deep_copy(a, b);
+void deepcopy_view(ViewTypeA& a, ViewTypeB& b, benchmark::State& state) {
+  const auto R = state.range(1);
+
+  for (auto _ : state) {
+    Kokkos::Timer timer;
+    for (int i = 0; i < R; i++) {
+      Kokkos::deep_copy(a, b);
+    }
+    Kokkos::fence();
+
+    report_results(state, timer.seconds() / R);
   }
-  Kokkos::fence();
-  return timer.seconds();
 }
 
 template <class LayoutA, class LayoutB>
-void run_deepcopyview_tests123(int N, int R) {
-  const int N1 = N;
+static void ViewDeepCopy_Rank1(benchmark::State& state) {
+  const int N8 = std::pow(state.range(0), 8);
+
+  Kokkos::View<double*, LayoutA> a("A1", N8);
+  Kokkos::View<double*, LayoutB> b("B1", N8);
+
+  deepcopy_view(a, b, state);
+}
+
+template <class LayoutA, class LayoutB>
+static void ViewDeepCopy_Rank2(benchmark::State& state) {
+  const int N1 = state.range(0);
+  const int N2 = N1 * N1;
+  const int N4 = N2 * N2;
+
+  Kokkos::View<double**, LayoutA> a("A2", N4, N4);
+  Kokkos::View<double**, LayoutB> b("B2", N4, N4);
+
+  deepcopy_view(a, b, state);
+}
+
+template <class LayoutA, class LayoutB>
+static void ViewDeepCopy_Rank3(benchmark::State& state) {
+  const int N1 = state.range(0);
   const int N2 = N1 * N1;
   const int N3 = N2 * N1;
-  const int N4 = N2 * N2;
-  const int N8 = N4 * N4;
 
-  double time1, time2, time3, time_raw = 100000.0;
-  {
-    Kokkos::View<double*, LayoutA> a("A1", N8);
-    Kokkos::View<double*, LayoutB> b("B1", N8);
-    time1 = deepcopy_view(a, b, R) / R;
-  }
-  {
-    Kokkos::View<double**, LayoutA> a("A2", N4, N4);
-    Kokkos::View<double**, LayoutB> b("B2", N4, N4);
-    time2 = deepcopy_view(a, b, R) / R;
-  }
-  {
-    Kokkos::View<double***, LayoutA> a("A3", N3, N3, N2);
-    Kokkos::View<double***, LayoutB> b("B3", N3, N3, N2);
-    time3 = deepcopy_view(a, b, R) / R;
-  }
-#if defined(KOKKOS_ENABLE_CUDA_LAMBDA) || !defined(KOKKOS_ENABLE_CUDA)
-  {
-    Kokkos::View<double*, LayoutA> a("A1", N8);
-    Kokkos::View<double*, LayoutB> b("B1", N8);
-    double* const a_ptr       = a.data();
-    const double* const b_ptr = b.data();
+  Kokkos::View<double***, LayoutA> a("A3", N3, N3, N2);
+  Kokkos::View<double***, LayoutB> b("B3", N3, N3, N2);
+
+  deepcopy_view(a, b, state);
+}
+
+template <class LayoutA, class LayoutB>
+static void ViewDeepCopy_Rank4(benchmark::State& state) {
+  const int N1 = state.range(0);
+  const int N2 = N1 * N1;
+
+  Kokkos::View<double****, LayoutA> a("A4", N2, N2, N2, N2);
+  Kokkos::View<double****, LayoutB> b("B4", N2, N2, N2, N2);
+
+  deepcopy_view(a, b, state);
+}
+
+template <class LayoutA, class LayoutB>
+static void ViewDeepCopy_Rank5(benchmark::State& state) {
+  const int N1 = state.range(0);
+  const int N2 = N1 * N1;
+
+  Kokkos::View<double*****, LayoutA> a("A5", N2, N2, N1, N1, N2);
+  Kokkos::View<double*****, LayoutB> b("B5", N2, N2, N1, N1, N2);
+
+  deepcopy_view(a, b, state);
+}
+
+template <class LayoutA, class LayoutB>
+static void ViewDeepCopy_Rank6(benchmark::State& state) {
+  const int N1 = state.range(0);
+  const int N2 = N1 * N1;
+
+  Kokkos::View<double******, LayoutA> a("A6", N2, N1, N1, N1, N1, N2);
+  Kokkos::View<double******, LayoutB> b("B6", N2, N1, N1, N1, N1, N2);
+
+  deepcopy_view(a, b, state);
+}
+
+template <class LayoutA, class LayoutB>
+static void ViewDeepCopy_Rank7(benchmark::State& state) {
+  const int N1 = state.range(0);
+  const int N2 = N1 * N1;
+
+  Kokkos::View<double*******, LayoutA> a("A7", N2, N1, N1, N1, N1, N1, N1);
+  Kokkos::View<double*******, LayoutB> b("B7", N2, N1, N1, N1, N1, N1, N1);
+
+  deepcopy_view(a, b, state);
+}
+
+template <class LayoutA, class LayoutB>
+static void ViewDeepCopy_Rank8(benchmark::State& state) {
+  const int N1 = state.range(0);
+
+  Kokkos::View<double********, LayoutA> a("A8", N1, N1, N1, N1, N1, N1, N1, N1);
+  Kokkos::View<double********, LayoutB> b("B8", N1, N1, N1, N1, N1, N1, N1, N1);
+
+  deepcopy_view(a, b, state);
+}
+
+template <class LayoutA, class LayoutB>
+static void ViewDeepCopy_Raw(benchmark::State& state) {
+  const auto N8 = std::pow(state.range(0), 8);
+  const auto R  = state.range(1);
+
+  Kokkos::View<double*, LayoutA> a("A1", N8);
+  Kokkos::View<double*, LayoutB> b("B1", N8);
+  double* const a_ptr       = a.data();
+  const double* const b_ptr = b.data();
+
+  for (auto _ : state) {
     Kokkos::Timer timer;
     for (int r = 0; r < R; r++) {
       Kokkos::parallel_for(
           N8, KOKKOS_LAMBDA(const int& i) { a_ptr[i] = b_ptr[i]; });
     }
     Kokkos::fence();
-    time_raw = timer.seconds() / R;
+
+    report_results(state, timer.seconds() / R);
   }
-#endif
-  double size = 1.0 * N8 * 8 / 1024 / 1024;
-  printf("   Raw:   %lf s   %lf MB   %lf GB/s\n", time_raw, size,
-         2.0 * size / 1024 / time_raw);
-  printf("   Rank1: %lf s   %lf MB   %lf GB/s\n", time1, size,
-         2.0 * size / 1024 / time1);
-  printf("   Rank2: %lf s   %lf MB   %lf GB/s\n", time2, size,
-         2.0 * size / 1024 / time2);
-  printf("   Rank3: %lf s   %lf MB   %lf GB/s\n", time3, size,
-         2.0 * size / 1024 / time3);
 }
 
 }  // namespace Test
+
+#endif
