+++ core/perf_test/PerfTest_ViewCopy_a45.cpp
@@ -42,7 +42,7 @@
 //@HEADER
 */
 
-#include <Benchmark_ViewCopy.hpp>
+#include <PerfTest_ViewCopy.hpp>
 
 namespace Test {
