+++ core/perf_test/PerfTest_ViewCopy_c8.cpp
@@ -42,7 +42,7 @@
 //@HEADER
 */
 
-#include <Benchmark_ViewCopy.hpp>
+#include <PerfTest_ViewCopy.hpp>
 
 // DeepCopy Performance for LayoutLeft to LayoutRight
 namespace Test {
